requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
Jinja2>=3.1.2
openai>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-openai>=0.0.2
langchain-core>=0.1.0
langchain-community>=0.0.10
# GHL MCP integration dependencies
aiohttp>=3.9.0
# Optional: For semantic search functionality (app works without these)
faiss-cpu>=1.7.0
numpy>=1.24.0
